:package: Migrate project dependencies to Pipenv
This commit reworks the project dependencies into a Pipenv-generated
Pipfile. It also specifies specific hashes and versions with the
Pipfile.lock file. This is a precursor task, leading up to more serious
Python packaging on the app. But this makes it easier for me to work on
the project, and ideally, others as well.